// types/species.ts
export type Species = {
  base_happiness: number
  capture_rate: number
  color: {
    name: string
    url: string
  }
  egg_groups: [
    {
      name:
        | 'bug'
        | 'ditto'
        | 'dragon'
        | 'fairy'
        | 'flying'
        | 'ground'
        | 'humanshape'
        | 'indeterminate'
        | 'mineral'
        | 'monster'
        | 'no-eggs'
        | 'plant'
        | 'water1'
        | 'water2'
        | 'water3'
      url: string
    },
  ]
  evolution_chain: {
    url: string
  }
  evolves_from_species: {
    name: string
    url: string
  }
  flavor_text_entries: {
    flavor_text: string
    language: {
      name: string
      url: string
    }
    version: {
      name: string
      url: string
    }
  }[]
  form_descriptions: {
    description: string
    language: {
      name: string
      url: string
    }
  }[]
  forms_switchable: boolean
  gender_rate: number
  genera: {
    genus: string
    language: {
      name: string
      url: string
    }
  }[]
  generation: {
    name: string
    url: string
  }
  growth_rate: {
    name:
      | 'fast'
      | 'fast-then-very-slow'
      | 'medium'
      | 'medium-slow'
      | 'slow'
      | 'slow-then-very-fast'
    url: string
  }
  habitat: {
    name: string
    url: string
  } | null
  has_gender_differences: boolean
  hatch_counter: number
  id: number
  is_baby: boolean
  is_legendary: boolean
  is_mythical: boolean
  name: string
  names: {
    language: {
      name: string
      url: string
    }
    name: string
  }[]
  order: number
  pal_park_encounters: {
    area: {
      name: string
      url: string
    }
    base_score: number
    rate: number
  }[]
  pokedex_numbers: {
    entry_number: number
    pokedex: {
      name: string
      url: string
    }
  }[]
  shape: {
    name: string
    url: string
  }
  varieties: {
    is_default: boolean
    pokemon: {
      name: string
      url: string
    }
  }[]
}